requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pyyaml
python-multipart
pydantic
pydantic-settings
pdfplumber
